% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_Functions.R
\name{Classify}
\alias{Classify}
\title{Classify}
\usage{
Classify(TrainData, TrainCat, TestData, Dvec = NULL, W = NULL,
  gamma = 1e-05)
}
\arguments{
\item{TrainData}{A (n x p) numeric matrix without missing values consisting of n training samples each with p features.}

\item{TrainCat}{A vector of length n consisting of group labels of the n training samples in \code{TrainData}. Must consist of 1s and 2s.}

\item{TestData}{A (m x p) numeric matrix without missing values consisting of m training samples each with p features. The number of features must equal the number of features in \code{TrainData}.}

\item{Dvec}{An optinal (p x 1) vector used as the discriminant vector. Defualt value is NULL, as the function generates its own discriminant vector.}

\item{W}{An optinal (p x p) matrix used as the within-group covariance matrix. Defualt value is NULL, as the function generates its own covariance matrix.}

\item{gamma}{A numeric value for the stabilization amount gamma * I added to the covariance matrixed used in the LDA decision rule. Default amount is 1E-5. Cannot be negative.}
}
\description{
A function which implements full LDA on the supplied data.

Generates linear discriminant vector and class predictions for \code{TestData}.
}
\details{
Function for full LDA.
}
