% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_Functions.R
\name{subsetPredict}
\alias{subsetPredict}
\title{subsetPredict}
\usage{
subsetPredict(TrainData, TrainCat, TestData, m1, m2, gamma = 1e-05)
}
\arguments{
\item{TrainData}{A (n x p) numeric matrix without missing values consisting of n training samples each with p features.}

\item{TrainCat}{A vector of length n consisting of group labels of the n training samples in \code{TrainData}. Must consist of 1s and 2s.}

\item{TestData}{A (m x p) numeric matrix without missing values consisting of m training samples each with p features. The number of features must equal the number of features in \code{TrainData}.}

\item{m1}{The number of class 1 sub-samples. Must be a positive integer.}

\item{m2}{The number of class 2 sub-samples. Must be a positive integer.}

\item{gamma}{A numeric value for the stabilization amount gamma * I added to the covariance matrixed used in the LDA decision rule. Default amount is 1E-5. Cannot be negative.}
}
\description{
A function which implements sub-sampled LDA on the supplied data.

Generates the sub-sampled linear discriminant vector and class predictions for \code{TestData}.
}
\details{
Function for sub-sampled LDA.
}
